Redux [Refactoring Files Name/Structure]
