--- src/reactivity/redux/ReduxPageHandlerComp.tsx
@@ -1,5 +1,0 @@
-function ReduxPageHandlerComp(props: { pageTitle: string; onClick: any }) {
-  return <button onClick={props.onClick}> {props.pageTitle} </button>
-}
-
-export default ReduxPageHandlerComp
