+++ src/reactivity/redux/components/ReduxSectionVisibilityButtonComp.tsx
@@ -1,0 +1,5 @@
+function ReduxSectionVisibilityButtonComp(props: { pageTitle: string; onClick: any }) {
+  return <button onClick={props.onClick}> {props.pageTitle} </button>
+}
+
+export default ReduxSectionVisibilityButtonComp
